+++ components/Layout/BottomNavigation.tsx
@@ -4,35 +4,11 @@ import { useRouter } from "next/navigation";
 import React, { useEffect, useState } from "react";
 import path from "@/routes";
 import { HomeSVG, PencilSVG, UserSVG } from "../svg";
+import useHideOnScroll from "@/hooks/useHideOnScroll";
 
 const BottomNavigation: React.FC = () => {
   const router = useRouter();
-  const [isVisible, setIsVisible] = useState(true);
-  const [lastScrollY, setLastScrollY] = useState(0);
-
-  useEffect(() => {
-    const controlNavbar = () => {
-      if (typeof window !== "undefined") {
-        if (window.scrollY > lastScrollY) {
-          // 스크롤 내릴 때
-          setIsVisible(false);
-        } else {
-          // 스크롤 올릴 때
-          setIsVisible(true);
-        }
-        setLastScrollY(window.scrollY);
-      }
-    };
-
-    if (typeof window !== "undefined") {
-      window.addEventListener("scroll", controlNavbar);
-
-      // 클린업 함수
-      return () => {
-        window.removeEventListener("scroll", controlNavbar);
-      };
-    }
-  }, [lastScrollY]);
+  const [isVisible] = useHideOnScroll();
   return (
     <div
       className={`fixed bottom-0 left-0 z-50 w-full bg-white border-t border-gray-200 dark:bg-gray-700 dark:border-gray-600 transition-transform duration-300 ${
